Add second option to display several number of routes that the total distance is smallest
The implementation resarches all possibilities, hence it is optimal but very slow.
This brute algorithm is generating all permutations of Stirling number - S(n, k), where n is number of deliveries and k is numnber of desired routes.
I had to implement it with the assumption, that every from k routes is nonempty. Otherwise, the optimal number of routes would be often 1.

Routes are displayed in one of several colors from array 'colors' chosen randomly in JS to provide more functionality.

This solution is still good for some small cases and for upcoming new versions for testing.

Looking for a more optimal soluton.
Ideas:
	- Devide the city into smaller regions and then distribute just regions betweeen routes,
	- Search for the points that are farest from the centre of all points and then start to generate routes,
	- Come up with some greedy algorithm